requirements.txt
flask[async]==2.2.2
python-telegram-bot==13.14
requests==2.28.1
python-dotenv==0.21.0
gunicorn==20.1.0
redis==4.3.4
